Remove subclasses and move transformer list to kafka connector
