+++ backend/streams_explorer/core/extractor/default/generic.py
@@ -7,18 +7,13 @@ from streams_explorer.models.kafka_connector import (
 )
 
 
-class GenericSinkConnector(KafkaConnector):
-    def get_topics(self) -> List[str]:
-        return Extractor.split_topics(self.config.topics)
-
-
 class GenericSink(Extractor):
     def on_connector_info_parsing(
         self, info: dict, connector_name: str
     ) -> Optional[KafkaConnector]:
         if info["type"] == KafkaConnectorTypesEnum.SINK:
             config = info["config"]
-            return GenericSinkConnector(
+            return KafkaConnector(
                 name=connector_name,
                 config=config,
                 type=KafkaConnectorTypesEnum.SINK,
