+++ backend/streams_explorer/core/extractor/default/s3_sink.py
@@ -8,11 +8,6 @@ from streams_explorer.models.kafka_connector import (
 from streams_explorer.models.sink import Sink
 
 
-class S3SinkConnector(KafkaConnector):
-    def get_topics(self) -> List[str]:
-        return Extractor.split_topics(self.config.topics)
-
-
 class S3Sink(Extractor):
     def __init__(self):
         self.sinks: List[Sink] = []
@@ -32,7 +27,7 @@ class S3Sink(Extractor):
                         source=connector_name,
                     )
                 )
-            return S3SinkConnector(
+            return KafkaConnector(
                 name=connector_name,
                 config=config,
                 type=KafkaConnectorTypesEnum.SINK,
